bufferpool: get the true offsize and  calculate the correct bufferpool. [2/2]
PD#OTT-17118

Problem:
The  allocation of "buffer pool" uses the actual resolution of the video,
but some videos have a larger "crop" value, and the "buffer pool" allocated
by the actual resolution is not consistent with the driver resolution.

Solution:
struct align value use the gstreamer analysis ,it include the true offsize.
we can use it calculate the correct bufferpool.

Verify:
gst-launch-1.0

Change-Id: I9d0251c7de623300a8b87d2be3005e9f3aabadc3
Signed-off-by: yichen.li <yichen.li@amlogic.com>